+++ src/snowflake/snowpark/_internal/analyzer/select_statement.py
@@ -818,12 +818,12 @@ def parse_column_name(column: Expression, analyzer: "Analyzer") -> Optional[str]
             # some expressions converted to SQL text with extra preceeding and trailing spaces.
             # Snowflake SQL removes the spaces in the returned column names.
             # So we remove it at the client too.
-            return analyzer.analyze(column, {}, parse_local_name=True).strip(" ")
+            return analyzer.analyze(column, parse_local_name=True).strip(" ")
         if isinstance(column, UnresolvedAttribute):
             if not column.is_sql_text:
                 return column.name
         if isinstance(column, UnresolvedAlias):
-            return analyzer.analyze(column, {}, parse_local_name=True).strip(" ")
+            return analyzer.analyze(column, parse_local_name=True).strip(" ")
         if isinstance(column, Alias):
             return column.name
     # We can parse column name from a column's SQL expression in the future.
@@ -922,7 +922,7 @@ def initiate_column_states(
     column_states = ColumnStateDict()
     for attr in column_attrs:
         # review later. should use parse_column_name
-        name = analyzer.analyze(attr, {}, parse_local_name=True).strip(" ")
+        name = analyzer.analyze(attr, parse_local_name=True).strip(" ")
         column_states[name] = ColumnState(
             name,
             change_state=ColumnChangeState.UNCHANGED_EXP,
@@ -988,7 +988,7 @@ def derive_column_states_from_subquery(
         from_c_state = from_.column_states.get(quoted_c_name)
         if from_c_state and from_c_state.change_state != ColumnChangeState.DROPPED:
             # review later. should use parse_column_name
-            if c_name != analyzer.analyze(c, {}, parse_local_name=True).strip(" "):
+            if c_name != analyzer.analyze(c, parse_local_name=True).strip(" "):
                 column_states[quoted_c_name] = ColumnState(
                     quoted_c_name,
                     ColumnChangeState.CHANGED_EXP,
