+++ src/snowflake/snowpark/_internal/udf_utils.py
@@ -766,7 +766,7 @@ def generate_call_python_sp_sql(
     sql_args = []
     for arg in args:
         if isinstance(arg, snowflake.snowpark.Column):
-            sql_args.append(session._analyzer.analyze(arg._expression, "CHANGEME"))
+            sql_args.append(session._analyzer.analyze(arg._expression))
         else:
             sql_args.append(to_sql(arg, infer_type(arg)))
     return f"CALL {sproc_name}({', '.join(sql_args)})"
